Stochastic Melodies v01 original upload
This patch follows the idea of generative stochastic music. It features 3 voices, of which 2 are sequencer-based and can be run in free or looped mode, and 1 is a solo voice that can be triggered manually. The notes played are inferred from probabilities for pitch, octave and duration set by the user. For each voice, cv outputs are an ADSR envelope (gate) and a pitch. An additional output is provided as a combination of the first voice and the solo voice. Here the solo voice overrides the first voice, allowing to manually perform on the first voice and create instant variations. See pdf for more details.